removed ListStops for each direction
